Review request 11648 - using entity proxy instead of account proxy. Also added LpiPMAccountExtensionImplDeploymentTestNG to test pm acc ext API with real data

+++ le-playmaker/src/main/java/com/latticeengines/playmaker/dao/PlaymakerRecommendationDao.java
@@ -10,42 +10,42 @@ public interface PlaymakerRecommendationDao extends GenericDao {
     List<Map<String, Object>> getRecommendations(long start, int offset, int maximum, int syncDestination,
             List<String> playIds);
 
-    int getRecommendationCount(long start, int syncDestination, List<String> playIds);
+    long getRecommendationCount(long start, int syncDestination, List<String> playIds);
 
     List<Map<String, Object>> getPlays(long start, int offset, int maximum, List<Integer> playgroupIds);
 
-    int getPlayCount(long start, List<Integer> playgroupIds);
+    long getPlayCount(long start, List<Integer> playgroupIds);
 
-    List<Map<String, Object>> getAccountExtensions(long start, int offset, int maximum, List<String> accountIds,
+    List<Map<String, Object>> getAccountExtensions(Long start, int offset, int maximum, List<String> accountIds,
             String filterBy, Long recStart, String columns, boolean hasSfdcContactId);
 
-    int getAccountExtensionCount(long start, List<String> accountIds, String filterBy, Long recStart);
+    long getAccountExtensionCount(Long start, List<String> accountIds, String filterBy, Long recStart);
 
     List<Map<String, Object>> getAccountExtensionSchema();
 
-    int getAccountExtensionColumnCount();
+    long getAccountExtensionColumnCount();
 
     List<Map<String, Object>> getPlayValues(long start, int offset, int maximum, List<Integer> playgroupIds);
 
-    int getPlayValueCount(long start, List<Integer> playgroupIds);
+    long getPlayValueCount(long start, List<Integer> playgroupIds);
 
     List<Map<String, Object>> getWorkflowTypes();
 
     List<Map<String, Object>> getPlayGroups(long start, int offset, int maximum);
 
-    int getPlayGroupCount(long start);
+    long getPlayGroupCount(long start);
 
     List<Map<String, Object>> getContacts(long start, int offset, int maximum, List<Integer> contactIds,
             List<Integer> accountIds);
 
-    int getContactCount(long start, List<Integer> contactIds, List<Integer> accountIds);
+    long getContactCount(long start, List<Integer> contactIds, List<Integer> accountIds);
 
     List<Map<String, Object>> getContactExtensionSchema();
 
-    int getContactExtensionCount(long start, List<Integer> contactIds);
+    long getContactExtensionCount(long start, List<Integer> contactIds);
 
     List<Map<String, Object>> getContactExtensions(long start, int offset, int maximum, List<Integer> contactIds);
 
-    int getContactExtensionColumnCount();
+    long getContactExtensionColumnCount();
 
 }
